Changed PKs to Integer, and fixed some verification logic.

+++ TcManagerWebApp/src/main/java/com/davidlowe/tcmanagerwebapp/services/AddressService.java
@@ -5,7 +5,7 @@ import com.davidlowe.tcmanagerwebapp.models.Address;
 import com.davidlowe.tcmanagerwebapp.models.Person;
 
 
-public interface AddressService extends CrudService<Address, Long>
+public interface AddressService extends CrudService<Address, Integer>
 {
     void deleteAddressForPerson(Person person);
 }
