+++ TcManagerWebApp/src/main/java/com/davidlowe/tcmanagerwebapp/services/BrokerageService.java
@@ -4,6 +4,6 @@ package com.davidlowe.tcmanagerwebapp.services;
 import com.davidlowe.tcmanagerwebapp.models.Brokerage;
 
 
-public interface BrokerageService extends CrudService<Brokerage, Long>
+public interface BrokerageService extends CrudService<Brokerage, Integer>
 {
 }
