+++ TcManagerWebApp/src/main/java/com/davidlowe/tcmanagerwebapp/services/DataInputStatusService.java
@@ -4,6 +4,6 @@ package com.davidlowe.tcmanagerwebapp.services;
 import com.davidlowe.tcmanagerwebapp.models.DataInputStatus;
 
 
-public interface DataInputStatusService extends ReadOnlyService<DataInputStatus, Long>
+public interface DataInputStatusService extends ReadOnlyService<DataInputStatus, Integer>
 {
 }
