+++ TcManagerWebApp/src/main/java/com/davidlowe/tcmanagerwebapp/services/PropertyTypeService.java
@@ -4,6 +4,6 @@ package com.davidlowe.tcmanagerwebapp.services;
 import com.davidlowe.tcmanagerwebapp.models.PropertyType;
 
 
-public interface PropertyTypeService extends ReadOnlyService<PropertyType, Long>
+public interface PropertyTypeService extends ReadOnlyService<PropertyType, Integer>
 {
 }
