+++ TcManagerWebApp/src/main/java/com/davidlowe/tcmanagerwebapp/services/TeamService.java
@@ -4,6 +4,6 @@ package com.davidlowe.tcmanagerwebapp.services;
 import com.davidlowe.tcmanagerwebapp.models.Team;
 
 
-public interface TeamService extends CrudService<Team, Long>
+public interface TeamService extends CrudService<Team, Integer>
 {
 }
